requirements.txt
flask==3.0.0
ultralytics==8.2.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
pandas==2.2.0
fpdf2==2.7.0